Clean Architecture + go_router
