+++ lib/config/router/app_router.dart
@@ -1,0 +1,13 @@
+import 'package:go_router/go_router.dart';
+import 'package:printer3d_cam/presentation/screens/camera/home_screen.dart';
+
+final appRouter = GoRouter(
+  initialLocation: '/',
+  routes: [
+    GoRoute(
+      path: '/',
+      name: HomeScreen.name,
+      builder: (context, state) => const HomeScreen(),
+    ),
+  ],
+);
