+++ lib/config/theme/app_theme.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class AppTheme {
+  ThemeData getTheme() => ThemeData(
+        useMaterial3: true,
+        colorSchemeSeed: const Color(0xff2862f5),
+      );
+}
