+++ pubspec.yaml
@@ -33,6 +33,7 @@ dependencies:
     sdk: flutter
   flutter_hooks: ^0.20.1
   flutter_mjpeg: ^2.0.4
+  go_router: ^10.1.2
 
 dev_dependencies:
   flutter_lints: ^2.0.0
